+++ app/src/main/java/com/example/eventradar/data/entities/Account.kt
@@ -1,10 +1,10 @@
-package com.example.eventradar.data
+package com.example.eventradar.data.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity
+@Entity(tableName = "account")
 data class Account(
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "e_mail") val eMail: String,
