+++ app/src/main/java/com/example/eventradar/data/entities/Address.kt
@@ -1,10 +1,10 @@
-package com.example.eventradar.data
+package com.example.eventradar.data.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity
+@Entity(tableName = "address")
 data class Address(
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "street") val street: Long,
