+++ app/src/main/java/com/example/eventradar/data/entities/Event.kt
@@ -1,10 +1,10 @@
-package com.example.eventradar.data
+package com.example.eventradar.data.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity
+@Entity(tableName = "event")
 data class Event(
     @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = "organizer_id") val organizerId: Long,
