+++ app/src/main/java/com/example/eventradar/data/entities/OrganizerWithAccount.kt
@@ -1,12 +1,12 @@
-package com.example.eventradar.data
+package com.example.eventradar.data.entities
 
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.Relation
 
 @Entity
-data class UserWithAccount(
-    @Embedded val user: User,
+data class OrganizerWithAccount(
+    @Embedded val organizer: Organizer,
     @Relation(
         parentColumn = "id",
         entityColumn = "account_id"
