+++ app/src/main/java/com/example/eventradar/data/entities/UserWithAccount.kt
@@ -1,12 +1,12 @@
-package com.example.eventradar.data
+package com.example.eventradar.data.entities
 
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.Relation
 
 @Entity
-data class OrganizerWithAccount(
-    @Embedded val organizer: Organizer,
+data class UserWithAccount(
+    @Embedded val user: User,
     @Relation(
         parentColumn = "id",
         entityColumn = "account_id"
