+++ app/src/main/java/com/example/eventradar/data/entities/ZipCode.kt
@@ -1,10 +1,10 @@
-package com.example.eventradar.data
+package com.example.eventradar.data.entities
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity
+@Entity(tableName = "zip_code")
 data class ZipCode(
     @PrimaryKey @ColumnInfo(name = "zip_code") val zipCode: String,
     @ColumnInfo(name = "city") val city: String
